Add the input daemon and the Sovereign compositor
Phase 4, Milestones 3 and 4, and the last of the roadmap. Both run in
Ring 3 with no privileged access to anything.

The compositor reaches the screen only through a shared buffer handle and
learns what to draw only through IPC. A client allocates a buffer, fills
it, shares it with the compositor and sends a present message; the
compositor maps the same physical frames and blits them into the scanout.
Nothing is copied through the kernel and nothing in Ring 0 knows what a
window is.

The input daemon owns console input and is the only thing holding a read
capability on it, which is what stops any other process from watching the
keyboard -- the keylogging the PRD calls out. It drops control codes and
forwards only printable characters and newlines, so a consumer never sees
a byte it did not ask for. Escape shuts the whole display stack down, and
the request travels daemon to compositor over IPC without the kernel
involved.

Buffers take their pixel depth from the display rather than assuming 32
bits. QEMU's framebuffer is 24-bit, and a client rendering four-byte
pixels into a three-byte scanout produces an image sheared a little
further right on every row -- a bug that looks like a corrupted blit rather
than a format mismatch.

The tests read the display's memory back and assert on the actual colour
at the requested coordinates, and that just past the surface's edge
nothing was touched. A blit that runs one row long or lands at the wrong
offset satisfies every structural check and fails these.

+++ kernel/src/pci.rs
@@ -257,7 +257,7 @@ pub fn read_bar(address: Address, index: u8) -> Option<Bar> {
                 (!combined_mask).wrapping_add(1),
             )
         } else {
-            let size = (!(mask & 0xFFFF_FFF0) as u32).wrapping_add(1) as u64;
+            let size = (!(mask & 0xFFFF_FFF0)).wrapping_add(1) as u64;
             (low, size)
         };
